Enforce architecture checks in CI
