+++ crates/apps/reticulumd/src/bin/reticulumd/rpc_loop_parts/read_http_request.rs
@@ -83,159 +83,11 @@ fn request_read_error_response(error: &io::Error) -> &'static [u8] {
     }
 }
 
-fn build_tls_server_config(config: &RpcTlsConfig) -> io::Result<std::sync::Arc<ServerConfig>> {
-    let server_chain = load_cert_chain(config.cert_chain_path.as_path())?;
-    let private_key = load_private_key(config.private_key_path.as_path())?;
-
-    let builder = ServerConfig::builder();
-    let server_config = if let Some(client_ca_path) = config.client_ca_path.as_ref() {
-        let roots = load_root_store(client_ca_path.as_path())?;
-        let verifier =
-            WebPkiClientVerifier::builder(std::sync::Arc::new(roots)).build().map_err(|err| {
-                io::Error::new(
-                    io::ErrorKind::InvalidInput,
-                    format!(
-                        "failed to build client verifier from {}: {}",
-                        client_ca_path.display(),
-                        err
-                    ),
-                )
-            })?;
-        builder
-            .with_client_cert_verifier(verifier)
-            .with_single_cert(server_chain, private_key)
-            .map_err(|err| {
-                io::Error::new(
-                    io::ErrorKind::InvalidInput,
-                    format!("invalid rpc tls server certificate/key configuration: {}", err),
-                )
-            })?
-    } else {
-        builder.with_no_client_auth().with_single_cert(server_chain, private_key).map_err(
-            |err| {
-                io::Error::new(
-                    io::ErrorKind::InvalidInput,
-                    format!("invalid rpc tls server certificate/key configuration: {}", err),
-                )
-            },
-        )?
-    };
-
-    Ok(std::sync::Arc::new(server_config))
-}
-
-fn load_cert_chain(path: &Path) -> io::Result<Vec<rustls::pki_types::CertificateDer<'static>>> {
-    let file = File::open(path)?;
-    let certificates = rustls::pki_types::CertificateDer::pem_reader_iter(file)
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(|err| {
-            io::Error::new(
-                io::ErrorKind::InvalidData,
-                format!("failed to parse PEM certs from {}: {}", path.display(), err),
-            )
-        })?;
-    if certificates.is_empty() {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("no certificates found in {}", path.display()),
-        ));
-    }
-    Ok(certificates)
-}
-
-fn load_private_key(path: &Path) -> io::Result<rustls::pki_types::PrivateKeyDer<'static>> {
-    let file = File::open(path)?;
-    let key = rustls::pki_types::PrivateKeyDer::pem_reader_iter(file)
-        .next()
-        .transpose()
-        .map_err(|err| {
-            io::Error::new(
-                io::ErrorKind::InvalidData,
-                format!("failed to parse private key {}: {}", path.display(), err),
-            )
-        })?;
-    key.ok_or_else(|| {
-        io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("no private key found in {}", path.display()),
-        )
-    })
-}
-
-fn load_root_store(path: &Path) -> io::Result<RootCertStore> {
-    let certificates = load_cert_chain(path)?;
-    let mut roots = RootCertStore::empty();
-    let (added, _ignored) = roots.add_parsable_certificates(certificates);
-    if added == 0 {
-        return Err(io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("no valid CA certificates found in {}", path.display()),
-        ));
-    }
-    Ok(roots)
-}
-
-fn extract_transport_auth(stream: &TlsStream<TcpStream>) -> http::TransportAuthContext {
-    let mut context = http::TransportAuthContext::default();
-    let (_tcp_stream, session) = stream.get_ref();
-    let Some(peer_certs) = session.peer_certificates() else {
-        return context;
-    };
-    let Some(leaf) = peer_certs.first() else {
-        return context;
-    };
-    context.client_cert_present = true;
-    let (subject, sans) = parse_client_identity(leaf.as_ref());
-    context.client_subject = subject;
-    context.client_sans = sans;
-    context
-}
-
-fn parse_client_identity(cert_der: &[u8]) -> (Option<String>, Vec<String>) {
-    let Ok((_remaining, cert)) = X509Certificate::from_der(cert_der) else {
-        return (None, Vec::new());
-    };
-    let subject = cert
-        .subject()
-        .iter_common_name()
-        .find_map(|name| name.as_str().ok().map(str::to_string))
-        .or_else(|| Some(cert.subject().to_string()));
-    let sans = parse_subject_alt_names(&cert);
-    (subject, sans)
-}
-
-fn parse_subject_alt_names(cert: &X509Certificate<'_>) -> Vec<String> {
-    let mut sans = Vec::new();
-    for extension in cert.extensions() {
-        if let ParsedExtension::SubjectAlternativeName(subject_alt_name) =
-            extension.parsed_extension()
-        {
-            for name in &subject_alt_name.general_names {
-                let value = match name {
-                    GeneralName::DNSName(value) => Some((*value).to_string()),
-                    GeneralName::URI(value) => Some((*value).to_string()),
-                    GeneralName::RFC822Name(value) => Some((*value).to_string()),
-                    GeneralName::IPAddress(raw) if raw.len() == 4 => {
-                        Some(IpAddr::from([raw[0], raw[1], raw[2], raw[3]]).to_string())
-                    }
-                    GeneralName::IPAddress(raw) if raw.len() == 16 => {
-                        let mut octets = [0_u8; 16];
-                        octets.copy_from_slice(raw);
-                        Some(IpAddr::from(octets).to_string())
-                    }
-                    _ => None,
-                };
-                if let Some(value) = value {
-                    let value = value.trim();
-                    if !value.is_empty() {
-                        sans.push(value.to_string());
-                    }
-                }
-            }
-        }
-    }
-    sans
-}
+#[path = "read_http_request_tls.rs"]
+mod read_http_request_tls;
+use read_http_request_tls::{build_tls_server_config, extract_transport_auth};
+#[cfg(test)]
+use read_http_request_tls::{load_cert_chain, load_private_key};
 
 #[cfg(test)]
 mod rpc_loop_tests {
