+++ crates/libs/rns-transport/src/packet_flags.rs
@@ -1,0 +1,173 @@
+use super::*;
+
+pub const LXMF_MAX_PAYLOAD: usize = PACKET_MDU - FERNET_OVERHEAD_SIZE - FERNET_MAX_PADDING_SIZE;
+pub const PACKET_IFAC_MAX_LENGTH: usize = 64usize;
+pub const PACKET_DATA_MAX: usize = 262_144usize;
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum IfacFlag {
+    Open = 0b0,
+    Authenticated = 0b1,
+}
+
+impl From<u8> for IfacFlag {
+    fn from(value: u8) -> Self {
+        match value {
+            0 => IfacFlag::Open,
+            1 => IfacFlag::Authenticated,
+            _ => IfacFlag::Open,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum HeaderType {
+    Type1 = 0b0,
+    Type2 = 0b1,
+}
+
+impl From<u8> for HeaderType {
+    fn from(value: u8) -> Self {
+        match value & 0b1 {
+            0 => HeaderType::Type1,
+            1 => HeaderType::Type2,
+            _ => HeaderType::Type1,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum PropagationType {
+    Broadcast = 0b0,
+    Transport = 0b1,
+}
+
+impl From<u8> for PropagationType {
+    fn from(value: u8) -> Self {
+        match value & 0b1 {
+            0b0 => PropagationType::Broadcast,
+            0b1 => PropagationType::Transport,
+            _ => PropagationType::Broadcast,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum ContextFlag {
+    Unset = 0b0,
+    Set = 0b1,
+}
+
+impl From<u8> for ContextFlag {
+    fn from(value: u8) -> Self {
+        match value & 0b1 {
+            0b0 => ContextFlag::Unset,
+            0b1 => ContextFlag::Set,
+            _ => ContextFlag::Unset,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum DestinationType {
+    Single = 0b00,
+    Group = 0b01,
+    Plain = 0b10,
+    Link = 0b11,
+}
+
+impl From<u8> for DestinationType {
+    fn from(value: u8) -> Self {
+        match value & 0b11 {
+            0b00 => DestinationType::Single,
+            0b01 => DestinationType::Group,
+            0b10 => DestinationType::Plain,
+            0b11 => DestinationType::Link,
+            _ => DestinationType::Single,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum PacketType {
+    Data = 0b00,
+    Announce = 0b01,
+    LinkRequest = 0b10,
+    Proof = 0b11,
+}
+
+impl From<u8> for PacketType {
+    fn from(value: u8) -> Self {
+        match value & 0b11 {
+            0b00 => PacketType::Data,
+            0b01 => PacketType::Announce,
+            0b10 => PacketType::LinkRequest,
+            0b11 => PacketType::Proof,
+            _ => PacketType::Data,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub enum PacketContext {
+    None = 0x00,                    // Generic data packet
+    Resource = 0x01,                // Packet is part of a resource
+    ResourceAdvrtisement = 0x02,    // Packet is a resource advertisement
+    ResourceRequest = 0x03,         // Packet is a resource part request
+    ResourceHashUpdate = 0x04,      // Packet is a resource hashmap update
+    ResourceProof = 0x05,           // Packet is a resource proof
+    ResourceInitiatorCancel = 0x06, // Packet is a resource initiator cancel message
+    ResourceReceiverCancel = 0x07,  // Packet is a resource receiver cancel message
+    CacheRequest = 0x08,            // Packet is a cache request
+    Request = 0x09,                 // Packet is a request
+    Response = 0x0A,                // Packet is a response to a request
+    PathResponse = 0x0B,            // Packet is a response to a path request
+    Command = 0x0C,                 // Packet is a command
+    CommandStatus = 0x0D,           // Packet is a status of an executed command
+    Channel = 0x0E,                 // Packet contains link channel data
+    KeepAlive = 0xFA,               // Packet is a keepalive packet
+    LinkIdentify = 0xFB,            // Packet is a link peer identification proof
+    LinkClose = 0xFC,               // Packet is a link close message
+    LinkProof = 0xFD,               // Packet is a link packet proof
+    LinkRTT = 0xFE,                 // Packet is a link request round-trip time measurement
+    LinkRequestProof = 0xFF,        // Packet is a link request proof
+}
+
+impl From<u8> for PacketContext {
+    fn from(value: u8) -> Self {
+        match value {
+            0x01 => PacketContext::Resource,
+            0x02 => PacketContext::ResourceAdvrtisement,
+            0x03 => PacketContext::ResourceRequest,
+            0x04 => PacketContext::ResourceHashUpdate,
+            0x05 => PacketContext::ResourceProof,
+            0x06 => PacketContext::ResourceInitiatorCancel,
+            0x07 => PacketContext::ResourceReceiverCancel,
+            0x08 => PacketContext::CacheRequest,
+            0x09 => PacketContext::Request,
+            0x0A => PacketContext::Response,
+            0x0B => PacketContext::PathResponse,
+            0x0C => PacketContext::Command,
+            0x0D => PacketContext::CommandStatus,
+            0x0E => PacketContext::Channel,
+            0xFA => PacketContext::KeepAlive,
+            0xFB => PacketContext::LinkIdentify,
+            0xFC => PacketContext::LinkClose,
+            0xFD => PacketContext::LinkProof,
+            0xFE => PacketContext::LinkRTT,
+            0xFF => PacketContext::LinkRequestProof,
+            _ => PacketContext::None,
+        }
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Copy, Clone)]
+pub struct Header {
+    pub ifac_flag: IfacFlag,
+    pub header_type: HeaderType,
+    pub context_flag: ContextFlag,
+    pub propagation_type: PropagationType,
+    pub destination_type: DestinationType,
+    pub packet_type: PacketType,
+    pub hops: u8,
+}
